+++ include/exadg/utilities/spatial_aware_function.h
